refact: rewrite update page using getServerSideProps
